backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
ldap3==2.9.1
psutil==5.9.8
aiofiles==23.2.1
python-dotenv==1.0.1
